Organiser/requirements.txt
-e git+https://github.com/django-nonrel/django@3d67d256ddf2ee1e602d3bb52ffee652ceaf6f5c#egg=Django-nonrel
-e git+https://github.com/sandlbn/django-bootstrap-calendar@3511353fe761212f82a80802dec99db574200409#egg=django_bootstrap_calendar-master
django-bootstrap3==4.7.0
-e git+https://github.com/django-nonrel/mongodb-engine@f01404715bc0503153b983af6d8a950eb2baeb78#egg=django_mongodb_engine-master
-e git+https://github.com/django-nonrel/djangotoolbox@6536b0e6080dce2ad89c22fa4ea671e1f7e817f5#egg=djangotoolbox-master
pymongo==2.7.1
wsgiref==0.1.2
